+++ client/js/Core/componentsandbox.js
@@ -21,6 +21,8 @@ SCWeb.core.ComponentSandbox = function (options) {
     this.is_struct = options.is_struct;
     this.format_addr = options.format_addr;
     this.is_editor = options.canEdit;
+    this.isRrelKeyScElement = false;
+    this.mainElement = null;
 
     this.eventGetObjectsToTranslate = null;
     this.eventApplyTranslation = null;
@@ -67,16 +69,16 @@ SCWeb.core.ComponentSandbox = function (options) {
         let addArcEventRequest = new sc.ScEventParams(
             new sc.ScAddr(this.addr),
             sc.ScEventType.AddOutgoingEdge,
-            (elAddr, edge, otherAddr) => {
-                if (self.eventStructUpdate) {
+            async (elAddr, edge, otherAddr) => {
+                if (self.eventStructUpdate && ((await scClient.checkElements([edge]))[0].equal(sc.ScType.EdgeAccessConstPosPerm))) {
                     self.eventStructUpdate(true, elAddr.value, edge.value);
                 }
             });
         let removeArcEventRequest = new sc.ScEventParams(
             new sc.ScAddr(this.addr),
             sc.ScEventType.RemoveOutgoingEdge,
-            (elAddr, edge, otherAddr) => {
-                if (self.eventStructUpdate) {
+            async (elAddr, edge, otherAddr) => {
+                if (self.eventStructUpdate && !(await window.scHelper.checkEdge(elAddr.value, sc.ScType.EdgeAccessConstPosPerm, otherAddr.value))) {
                     self.eventStructUpdate(false, elAddr.value, edge.value);
                 }
             });
@@ -261,16 +263,268 @@ SCWeb.core.ComponentSandbox.prototype.createViewersForScStructs = function (cont
  * {String} contentType type of content data (@see scClient.getLinkContent). If it's null, then
  * data will be returned as string
  */
-SCWeb.core.ComponentSandbox.prototype.updateContent = async function (scene, contentType) {
+SCWeb.core.ComponentSandbox.prototype.updateContent = async function (scAddr, scene, contentType) {
+    let edgeToEdge = false;
+    let relationNodes = [];    
     var self = this;
 
     if (scene) {
         self.scene = scene;
     } 
-
     if (this.is_struct && this.eventStructUpdate) {
         const maxNumberOfTriplets = 850;
         const delayTimeoutAutosize = 300;
+        const levelScale1 = { node: 2.3, link: 1.8, opacity: 1, widthEdge: 8, stroke: 'black', fill: '#00a' };
+        const levelScale2 = { node: 1.8, link: 1.5, opacity: 1, widthEdge: 7.5, stroke: 'black', fill: '#00a' };
+        const levelScale3 = { node: 1.4, link: 1, opacity: 1, widthEdge: 7, stroke: 'black', fill: '#00a' };
+        const levelScale4 = { node: 1, link: 1, opacity: 1, widthEdge: 6.5, stroke: 'black', fill: '#00a' };
+        const levelScale5 = { node: 1, link: 1, opacity: 0.8, widthEdge: 6.5, stroke: 'black', fill: '#00a' };
+        const levelScale6 = { node: 1, link: 1, opacity: 0.6, widthEdge: 6.5, stroke: 'black', fill: '#00a' };
+        const levelScale7 = { node: 1, link: 1, opacity: 0.4, widthEdge: 6.5, stroke: 'black', fill: '#00a' };
+
+        const levelScales = [levelScale1, levelScale2, levelScale3, levelScale4, levelScale5, levelScale6, levelScale7];
+        
+        const checkEdge = (scAddr) => {
+            let elem = self.scene.getObjectByScAddr(scAddr);
+            if (elem instanceof SCg.ModelEdge) return true;
+            return false;
+        };
+
+        let scTemplateMainlevel = new sc.ScTemplate();        
+        let scTemplateMainlevelWithMainKey = new sc.ScTemplate();   
+
+        if (scAddr && !self.isResultWithMainKey) {
+            scTemplateMainlevel.triple(
+                [new sc.ScAddr(self.addr), "src"],
+                [sc.ScType.EdgeAccessVarPosPerm, "edgeFromContourToMainNode"],
+                new sc.ScAddr(scAddr),
+            )
+        };
+        if (scAddr && self.isResultWithMainKey) {
+            scTemplateMainlevelWithMainKey.triple(
+                [new sc.ScAddr(self.addr), "src"],
+                [sc.ScType.EdgeAccessVarPosPerm, "edgeFromContourToMainNode"],
+                new sc.ScAddr(scAddr),
+            )
+        };
+        if (!scAddr) {
+            scTemplateMainlevel.tripleWithRelation(
+                [new sc.ScAddr(self.addr), "src"],
+                [sc.ScType.EdgeAccessVarPosPerm, "edgeFromContourToMainNode"],
+                [sc.ScType.Unknown, "mainNode"],
+                sc.ScType.EdgeAccessVarPosPerm,
+                new sc.ScAddr(window.scKeynodes['rrel_key_sc_element']),
+            );
+        };
+        if (!scAddr) {
+            scTemplateMainlevelWithMainKey.tripleWithRelation(
+                [new sc.ScAddr(self.addr), "src"],
+                [sc.ScType.EdgeAccessVarPosPerm, "edgeFromContourToMainNode"],
+                [sc.ScType.Unknown, "mainNode"],
+                sc.ScType.EdgeAccessVarPosPerm,
+                new sc.ScAddr(window.scKeynodes['rrel_main_key_sc_element']),
+            );
+        };
+        
+        let resultLevel = await window.scClient.templateSearch(scTemplateMainlevel);
+        let resultLevelWithMainKey = await window.scClient.templateSearch(scTemplateMainlevelWithMainKey);
+
+        let mainElements = [];
+
+        if (!scAddr) resultLevel.length ? self.isResultWithMainKey = false : isResultWithMainKey = true;
+
+        if (resultLevel.length || resultLevelWithMainKey.length) self.isRrelKeyScElement = true;
+        
+        for (let triple of resultLevel.length ? resultLevel : resultLevelWithMainKey) {
+
+            scAddr ? mainElements.push(scAddr) && (self.mainElement = scAddr): mainElements.push(triple.get('mainNode').value) && (self.mainElement = triple.get('mainNode').value);
+            self.eventStructUpdate(true, triple.get('src').value, triple.get('edgeFromContourToMainNode').value, levelScales[0]);
+        };
+
+        if (scene && !self.isResultWithMainKey) {
+            let isEdge = scAddr ? scene.getObjectByScAddr(scAddr).edges[0].target : scene.getObjectByScAddr(resultLevel.get('mainNode').value).edges[0].target;
+            isEdge instanceof SCg.ModelEdge ? edgeToEdge = true : edgeToEdge = false;
+        };
+        if (scene && self.isResultWithMainKey) {
+            let isEdge = scAddr ? scene.getObjectByScAddr(scAddr).edges[0].target : scene.getObjectByScAddr(resultLevelWithMainKey.get('mainNode').value).edges[0].target;
+            isEdge instanceof SCg.ModelEdge ? edgeToEdge = true : edgeToEdge = false;
+        };
+
+        let searchAllLevelEdges = async function (elementsArr, levelScales, level, visitedElements) {
+            let levelScale;
+
+            level > 6 ? levelScale = levelScale7 : levelScale = levelScales[level];
+
+            let newElementsArr = [];
+            for (let i = 0; i < elementsArr.length; i++) {
+                
+                let elements = elementsArr[i];
+                for (let j = 0; j < elements.length; j++) {
+                    let elem = elements[j];
+                    let newElements = await searchLevelEdges(elem, levelScale, visitedElements);
+                    if (newElements.length) newElementsArr.push(newElements);
+                }
+                await searchAllLevelEdges(newElementsArr, levelScales, level + 1, visitedElements);
+            }
+        };
+
+        let searchLevelEdges = async function (mainElem, scale, visitedElements) {
+            let outgoingLevelNodesWithRelation = await searchLevelEdgesByDirection(mainElem, scale, visitedElements, false, true);
+            let outgoingLevelNodesNotWithRelation = await searchLevelEdgesByDirection(mainElem, scale, visitedElements, false, false);
+            let incomingLevelNodesNotWithRelation = await searchLevelEdgesByDirection(mainElem, scale, visitedElements, true, false);
+            let incomingLevelNodesWithRelation = await searchLevelEdgesByDirection(mainElem, scale, visitedElements, true, true);
+            return [...incomingLevelNodesWithRelation, ...incomingLevelNodesNotWithRelation, ...outgoingLevelNodesWithRelation, ...outgoingLevelNodesNotWithRelation];
+        };
+
+        let searchLevelEdgesByDirection = async function (mainElem, scale, visitedElements, incomingEdge, withRelation) {
+            let levelNodes = [];
+
+            let scTemplateSearchEdgeElements = new sc.ScTemplate();
+
+            if (edgeToEdge) {
+                scTemplateSearchEdgeElements.triple(
+                    new sc.ScAddr(mainElem),
+                    sc.ScType.EdgeAccessVarPosPerm,
+                    [sc.ScType.Unknown, "secondElem"],
+                );
+
+                scTemplateSearchEdgeElements.triple(
+                    [new sc.ScAddr(self.addr), "src"],
+                    [sc.ScType.EdgeAccessVarPosPerm, "edgeFromContourToSecondElem"],
+                    "secondElem",
+                );
+
+                scTemplateSearchEdgeElements.triple(
+                    [sc.ScType.Unknown, "sourceElem"],
+                    "secondElem",
+                    [sc.ScType.Unknown, "targetElem"],
+                );
+            };
+
+            let resultEdgeElements = await window.scClient.templateSearch(scTemplateSearchEdgeElements);
+
+            let scTemplate = new sc.ScTemplate();
+            scTemplate.triple(
+                [new sc.ScAddr(self.addr), "src"],
+                [sc.ScType.EdgeAccessVarPosPerm, "edgeFromContourToMainEdge"],
+                [sc.ScType.Unknown, "edgeFromMainNodeToSecondElem"],
+            );
+
+            if (incomingEdge) {
+                scTemplate.triple(
+                    [sc.ScType.Unknown, "secondElem"],
+                    "edgeFromMainNodeToSecondElem",
+                    new sc.ScAddr(mainElem),
+                );
+            };
+            if (!incomingEdge) {
+                scTemplate.triple(
+                    new sc.ScAddr(mainElem),
+                    "edgeFromMainNodeToSecondElem",
+                    [sc.ScType.Unknown, "secondElem"],
+                );
+            };
+
+            scTemplate.triple(
+                "src",
+                [sc.ScType.EdgeAccessVarPosPerm, "edgeFromContourToSecondElem"],
+                "secondElem",
+            );
+
+            if (withRelation) {
+                scTemplate.triple(
+                    [sc.ScType.Unknown, "relationNode"],
+                    [sc.ScType.EdgeAccessVarPosPerm, "edgeFromRelationNodeToedgeFromMainNodeToSecondElem"],
+                    "edgeFromMainNodeToSecondElem",
+                );
+                scTemplate.triple(
+                    "src",
+                    [sc.ScType.EdgeAccessVarPosPerm, "edgeFromContourToRelationNode"],
+                    "relationNode",
+                );
+                scTemplate.triple(
+                    "src",
+                    [sc.ScType.EdgeAccessVarPosPerm, "edgeFromContourToEdgeFromRelationNodeToedgeFromMainNodeToSecondElem"],
+                    "edgeFromRelationNodeToedgeFromMainNodeToSecondElem",
+                );
+            };
+
+            let result = await window.scClient.templateSearch(scTemplate);
+            for (let triple of result) {
+                const secondElem = triple.get("secondElem").value;
+                if (checkEdge(secondElem) && !edgeToEdge) continue;
+
+                if (edgeToEdge) {
+                    relationNodes.push(mainElem);
+
+                    for (triple of resultEdgeElements) {
+                        const targetElem = triple.get("targetElem").value;
+                        const sourceElem = triple.get("sourceElem").value;
+
+                        let scTemplateSearchTargetNodes = new sc.ScTemplate();
+                        scTemplateSearchTargetNodes.triple(
+                            [new sc.ScAddr(self.addr), "src"],
+                            [sc.ScType.EdgeAccessVarPosPerm, "edgeFromContourToTargetElem"],
+                            new sc.ScAddr(targetElem),
+                        );
+
+                        scTemplateSearchTargetNodes.triple(
+                            "src",
+                            [sc.ScType.EdgeAccessVarPosPerm, "edgeFromContourToSourceElem"],
+                            new sc.ScAddr(sourceElem),
+                        );
+                        
+                        if (!visitedElements.includes(targetElem) && !levelNodes.includes(targetElem) || !visitedElements.includes(sourceElem) && !levelNodes.includes(sourceElem)) {
+                                
+                            levelNodes.push(targetElem);
+                            visitedElements.push(targetElem);
+                            levelNodes.push(sourceElem);
+                            visitedElements.push(sourceElem);
+
+                            self.eventStructUpdate(true, triple.get("src").value, triple.get("edgeFromContourToSecondElem").value, scale);
+                        };
+
+                        let resultEdgeElementsEdges = await window.scClient.templateSearch(scTemplateSearchTargetNodes);
+
+                        for (triple of resultEdgeElementsEdges) {
+                            self.eventStructUpdate(true, triple.get("src").value, triple.get("edgeFromContourToSourceElem").value, scale);
+                            self.eventStructUpdate(true, triple.get("src").value, triple.get("edgeFromContourToTargetElem").value, scale);
+                        };
+                    };
+                    edgeToEdge = false;
+                    continue;
+                };
+
+                if (!visitedElements.includes(secondElem) && !levelNodes.includes(secondElem)) {
+                    levelNodes.push(secondElem);
+                    visitedElements.push(secondElem);
+                    self.eventStructUpdate(true, triple.get("src").value, triple.get("edgeFromContourToSecondElem").value, scale);
+                    self.eventStructUpdate(true, triple.get("src").value, triple.get("edgeFromContourToMainEdge").value, scale);
+                };
+                
+                if (!levelNodes.length) continue;
+                if (secondElem === mainElements[0]) continue;
+
+                if (withRelation) {
+                    if (scAddr === mainElements[0] && edgeToEdge) continue;
+                    let relationNode = triple.get("relationNode").value;
+
+                    if (relationNodes.includes(relationNode)) continue;
+
+                    if (!visitedElements.includes(relationNode) && !levelNodes.includes(relationNode)) {
+                        levelNodes.push(relationNode);
+                        visitedElements.push(relationNode);
+                    }
+                    self.eventStructUpdate(true, triple.get("src").value, triple.get("edgeFromContourToRelationNode").value, scale);
+                    self.eventStructUpdate(true, triple.get("src").value, triple.get("edgeFromContourToEdgeFromRelationNodeToedgeFromMainNodeToSecondElem").value, scale);
+                };
+            };
+            return [...new Set(levelNodes)];
+        };
+
+        searchAllLevelEdges([mainElements], levelScales, 1, [...mainElements]);
+        if (mainElements.length) return;
+
         let scTemplate = new sc.ScTemplate();
         scTemplate.triple(
             [new sc.ScAddr(this.addr), "src"],
@@ -281,29 +535,8 @@ SCWeb.core.ComponentSandbox.prototype.updateContent = async function (scene, con
             result.splice(maxNumberOfTriplets-1, result.length-maxNumberOfTriplets);
         }
         for (let triple of result) {
-            self.eventStructUpdate(true, triple.get("src").value, triple.get("edge").value);
+            self.eventStructUpdate(true, triple.get("src").value, triple.get("edge").value, { node: 1, link: 1 });
         }
-
-        setTimeout(() => {
-            if (self.scene) {
-                const [conteinerWidth, conteinerHeight] = self.scene.getContainerSize();
-                const deltaTranslate = 0.1;
-                const addSize = 400;
-
-                const scgHeight = self.scene.render.d3_container[0][0].getBoundingClientRect().height + (conteinerHeight / 2 + addSize);
-                const scgWidth = self.scene.render.d3_container[0][0].getBoundingClientRect().width + (conteinerWidth / 2 + addSize);
-
-                const containerAspectRatio = conteinerHeight / conteinerWidth;
-                const graphAspectRatio = scgHeight / scgWidth;
-                
-                const scale = containerAspectRatio < graphAspectRatio ?
-                    conteinerWidth / scgWidth :
-                    conteinerHeight / scgHeight;
-                
-                
-                self.scene.render._changeContainerTransform([conteinerWidth * deltaTranslate, conteinerHeight * deltaTranslate], scale);
-            }
-        }, delayTimeoutAutosize);
     }
     else {
         let content = await window.scClient.getLinkContents([new sc.ScAddr(this.addr)]);