+++ scripts/set_vars.sh
@@ -1,0 +1,7 @@
+#!/usr/bin/env bash
+set -eo pipefail
+
+ROOT_PATH=$(cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && cd .. && pwd)
+
+export APP_ROOT_PATH="${APP_ROOT_PATH:-${ROOT_PATH}}"
+export SC_WEB_PATH="${SC_WEB_PATH:-${APP_ROOT_PATH}}"